reserved edit/delete for logged user

+++ controllers/events.js
@@ -4,13 +4,10 @@ function newEvent(req, res) {
     res.render('newEvent', { title: 'Create New Event' });
 }
 
-function newEvent(req, res) {
-    res.render('newEvent', { title: 'Create New Event' });
-};
-
 async function create(req, res) {
     try {
         const newVenue = new Venue(req.body);
+        newVenue.user = req.user._id;
         await newVenue.save();
         res.redirect('/'); 
     } catch (err) { }
@@ -18,30 +15,41 @@ async function create(req, res) {
 
 async function listEvents(req, res) {
     try {
-        const events = await Venue.find();
+        const events = await Venue.find().populate('user');
         res.render('index', { title: 'venue.Me', events: events });
     } catch (err) { }
 }
 
-    async function editEventForm(req, res) {
-        try {
-            const event = await Venue.findById(req.params.id);
-            res.render('editEvent', { event: event, title: 'Edit Event' });
-        } catch (err) { }
+async function editEventForm(req, res) {
+    try {
+        const event = await Venue.findById(req.params.id);
+        if (!event || !event.user.equals(req.user._id)) {
+            return;
+        }
+        res.render('editEvent', { event: event, title: 'Edit Event' });
+    } catch (err) { }
 }
-    
-    async function updateEvent(req, res) {
-        try {
-            await Venue.findByIdAndUpdate(req.params.id, req.body);
-            res.redirect('/');
-        } catch (err) { }
+
+async function updateEvent(req, res) {
+    try {
+        const event = await Venue.findById(req.params.id);
+        if (!event || !event.user.equals(req.user._id)) {
+            return; 
+        }
+        await Venue.findByIdAndUpdate(req.params.id, req.body);
+        res.redirect('/');
+    } catch (err) { }
 }
-    
-    async function deleteEvent(req, res) {
-        try {
-            await Venue.findByIdAndDelete(req.params.id);
-            res.redirect('/');
-        } catch (err) { } 
+
+async function deleteEvent(req, res) {
+    try {
+        const event = await Venue.findById(req.params.id);
+        if (!event || !event.user.equals(req.user._id)) {
+            return; 
+        }
+        await Venue.findByIdAndDelete(req.params.id);
+        res.redirect('/');
+    } catch (err) { }
 }
 
 module.exports = {
